backend/requirements.txt
# ============================================================================
# Backend Dependencies
# Khoa luan: Section 1.6.2 (Django REST), Section 4.1.1 (Moi truong)
# ============================================================================

# --- Django & REST API (Section 1.6.2) ---
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1

# --- Supabase Database ---
supabase==2.10.0

# --- AI Pipeline (Section 2.4.1) ---
# Buoc 1: Blur Detection (Section 1.5.2 - Laplacian Variance)
opencv-python-headless==4.9.0.80
numpy==1.26.4

# Buoc 2: Document Classification (Section 1.3 - EfficientNet-B0)
torch==2.1.2
torchvision==0.16.2

# Buoc 5: OCR (Section 1.4.2 - VietOCR)
vietocr==0.3.5
Pillow==10.2.0

# --- Utilities ---
python-dotenv==1.0.1
gunicorn==21.1.0
requests==2.31.0
